Improved matching, naming of preprocessor keywords
Added the defined preprocessor keywords (+include, see previous
commit for suggestion), instead of whatever kind of black magic
that was blocking theming of directive parameters. Now, only
a valid preprocessor keyword will be highlighted, and any literals
or symbols behind it will also be properly highlighted. Also
changed the scope name to hit better with common themes (Tested
with Microsoft themes, which colour preprocessor directives
differently, as well as with variants of Atom and Material and
others).

Signed-off-by: Andreas K. Berg <andreaskb98@gmail.com>